backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Config
python-decouple==3.8

# Image handling
Pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# CORS
python-cors==1.7.0

# File uploads
aiofiles==23.2.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring
sentry-sdk[fastapi]==1.38.0

# Rate limiting
slowapi==0.1.9

# Caching
aiocache==0.12.2

# Email (for notifications)
fastapi-mail==1.4.1

# WebSocket support
websockets==12.0

# Data validation
email-validator==2.1.0

# HTTP client for external APIs
httpx==0.25.2

# Environment types
types-passlib==1.7.7.13
types-python-jose==3.3.4.8